pyproject.toml
[project]
name = "secret-management-system"
description = "Secret Management System"
requires-python = ">=3.11"
version = "0.1.0"

dependencies = [
    "Flask==3.0.0",
    "Werkzeug==3.0.1",
    "cryptography==41.0.7",
    "argon2-cffi==23.1.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
    "google-api-python-client==2.111.0",
    "google-auth==2.25.2",
    "google-auth-oauthlib==1.2.0",
]